+++ packages/expo-updates/src/Updates.ts
@@ -1,25 +1,15 @@
-import {
-  DeviceEventEmitter,
-  CodedError,
-  NativeModulesProxy,
-  UnavailabilityError,
-} from 'expo-modules-core';
-import { EventEmitter, EventSubscription } from 'fbemitter';
+import { CodedError, NativeModulesProxy, UnavailabilityError } from 'expo-modules-core';
 
 import ExpoUpdates from './ExpoUpdates';
 import {
   LocalAssets,
   Manifest,
   UpdateCheckResult,
-  UpdateEvent,
   UpdateFetchResult,
   UpdatesCheckAutomaticallyValue,
   UpdatesLogEntry,
-  UpdatesNativeStateChangeEvent,
   UpdatesNativeStateMachineContext,
 } from './Updates.types';
-
-export * from './Updates.types';
 
 /**
  * The UUID that uniquely identifies the currently running update if `expo-updates` is enabled. The
@@ -298,82 +288,6 @@ export function clearUpdateCacheExperimentalAsync(_sdkVersion?: string) {
     "This method is no longer necessary. `expo-updates` now automatically deletes your app's old bundle files!"
   );
 }
-
-let _emitter: EventEmitter | null;
-
-function _getEmitter(): EventEmitter {
-  if (!_emitter) {
-    _emitter = new EventEmitter();
-    DeviceEventEmitter.addListener('Expo.nativeUpdatesEvent', _emitEvent);
-    DeviceEventEmitter.addListener(
-      'Expo.nativeUpdatesStateChangeEvent',
-      _emitNativeStateChangeEvent
-    );
-  }
-  return _emitter;
-}
-
-function _emitEvent(params): void {
-  let newParams = { ...params };
-  if (typeof params === 'string') {
-    newParams = JSON.parse(params);
-  }
-  if (newParams.manifestString) {
-    newParams.manifest = JSON.parse(newParams.manifestString);
-    delete newParams.manifestString;
-  }
-
-  if (!_emitter) {
-    throw new Error(`EventEmitter must be initialized to use from its listener`);
-  }
-  _emitter.emit('Expo.updatesEvent', newParams);
-}
-
-// Handle native state change events
-function _emitNativeStateChangeEvent(params: any) {
-  let newParams = { ...params };
-  if (typeof params === 'string') {
-    newParams = JSON.parse(params);
-  }
-  if (newParams.context.latestManifestString) {
-    newParams.context.latestManifest = JSON.parse(newParams.context.latestManifestString);
-    delete newParams.context.latestManifestString;
-  }
-  if (newParams.context.downloadedManifestString) {
-    newParams.context.downloadedManifest = JSON.parse(newParams.context.downloadedManifestString);
-    delete newParams.context.downloadedManifestString;
-  }
-  if (!_emitter) {
-    throw new Error(`EventEmitter must be initialized to use from its listener`);
-  }
-  _emitter?.emit('Expo.updatesStateChangeEvent', newParams);
-}
-
-/**
- * Adds a callback to be invoked when updates-related events occur (such as upon the initial app
- * load) due to auto-update settings chosen at build-time. See also the
- * [`useUpdateEvents`](#useupdateeventslistener) React hook.
- *
- * @param listener A function that will be invoked with an [`UpdateEvent`](#updateevent) instance
- * and should not return any value.
- * @return An `EventSubscription` object on which you can call `remove()` to unsubscribe the
- * listener.
- */
-export function addListener(listener: (event: UpdateEvent) => void): EventSubscription {
-  const emitter = _getEmitter();
-  return emitter.addListener('Expo.updatesEvent', listener);
-}
-
-/**
- * @hidden
- */
-export const addUpdatesStateChangeListener = (
-  listener: (event: UpdatesNativeStateChangeEvent) => void
-) => {
-  // Add listener for state change events
-  const emitter = _getEmitter();
-  return emitter.addListener('Expo.updatesStateChangeEvent', listener);
-};
 
 /**
  * @hidden