+++ packages/expo-updates/src/index.ts
@@ -1,2 +1,6 @@
 export * from './Updates';
+export * from './UpdatesEmitter';
 export * from './UpdatesHooks';
+export * from './Updates.types';
+export * from './UseUpdates';
+export * from './UseUpdates.types';